# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/marketplace/prompts": {
      "get": {
        "description": "Retrieves a paginated list of published prompts for the current tenant. Results can be filtered, sorted and searched.",
        "operationId": "MarketplaceController_getPublishedPrompts",
        "parameters": [
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search term to filter prompts by name or description",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number for pagination (starts at 1)",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of items per page",
            "schema": {
              "default": 10,
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "Field to sort by (e.g. name, createdAt)",
            "schema": {
              "default": "createdAt",
              "example": "createdAt",
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "description": "Sort order direction",
            "schema": {
              "enum": [
                "ASC",
                "DESC"
              ],
              "type": "string"
            }
          },
          {
            "name": "languageCode",
            "required": false,
            "in": "query",
            "description": "Language code to filter prompts (e.g., en-US, es-ES)",
            "schema": {
              "example": "en-US",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of published prompts retrieved successfully"
          },
          "400": {
            "description": "Invalid query parameters - Check the provided values"
          },
          "401": {
            "description": "Unauthorized - Invalid or expired token"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get published prompts",
        "tags": [
          "Marketplace"
        ]
      }
    },
    "/api/marketplace/assets": {
      "get": {
        "description": "Retrieves a paginated list of published assets for the current tenant. Results can be filtered, sorted and searched.",
        "operationId": "MarketplaceController_getPublishedAssets",
        "parameters": [
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search term to filter assets by name or description",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number for pagination (starts at 1)",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of items per page",
            "schema": {
              "default": 10,
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "Field to sort by (e.g. name, createdAt)",
            "schema": {
              "default": "createdAt",
              "example": "createdAt",
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "description": "Sort order direction",
            "schema": {
              "enum": [
                "ASC",
                "DESC"
              ],
              "type": "string"
            }
          },
          {
            "name": "languageCode",
            "required": false,
            "in": "query",
            "description": "Language code to filter assets (e.g., en-US, es-ES)",
            "schema": {
              "example": "en-US",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of published assets retrieved successfully"
          },
          "400": {
            "description": "Invalid query parameters - Check the provided values"
          },
          "401": {
            "description": "Unauthorized - Invalid or expired token"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get published assets",
        "tags": [
          "Marketplace"
        ]
      }
    }
  },
  "info": {
    "title": "japm.app API",
    "description": "API for the japm.app Prompt Engineering application",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
...
  }
}